Clean build

+++ gameCommon.hpp
@@ -40,8 +40,8 @@ typedef uint8_t byte;     // BYTE = unsigned 8 bit value
 // Fixed Point Math Helpers
 #define FP_SHIFT 16     // shifts to produce a fixed-point number
 #define FP_SCALE 65536  // scaling factor
-typedef int FIXPOINT;
-#define INT_TO_FIXP(n) (FIXPOINT((n << FP_SHIFT)))
+typedef int32_t FIXPOINT;
+#define INT_TO_FIXP(n) (FIXPOINT(((n) << FP_SHIFT)))
 #define FIXP_TO_INT(n) (int((n >> FP_SHIFT)))
 #define FLOAT_TO_FIXP(n) (FIXPOINT((float)n * FP_SCALE))
 #define FIXP_TO_FLOAT(n) (float((float)n / FP_SCALE))
@@ -49,7 +49,7 @@ typedef int FIXPOINT;
 #define FIXP_FIXP_INT_PART(n) (n & 0xffff0000)
 #define FIXP_DEC_PART(n) (n & 0x0000ffff)
 #define FIXP_MULT(n,n2) ((((int64_t)(n)) * n2) >> FP_SHIFT)
-#define FIXP_DIV(n,n2) (((((int64_t)(n)) << FP_SHIFT) / n2))
+#define FIXP_DIV(n,n2) (int32_t(((((int64_t)(n)) << FP_SHIFT) / n2)))
 #define FIXEDPT_PI FLOAT_TO_FIXP(PI)
 #define FIXEDPT_HALF_PI FLOAT_TO_FIXP(PI/2)
